+++ examples/resources/graviteeam_org_user/resource.tf
@@ -6,4 +6,7 @@ resource "graviteeam_org_user" "example" {
   last_name        = "User"
   enabled          = true
   pre_registration = true
+
+  reset_password         = "NewSecurePass123!"
+  reset_password_trigger = "rotation-2026-01"
 }