+++ examples/resources/graviteeam_user/resource.tf
@@ -6,4 +6,8 @@ resource "graviteeam_user" "example" {
   last_name        = "User"
   enabled          = true
   pre_registration = true
+
+  reset_password         = "SecurePass123!"
+  reset_password_trigger = "rotation-2026-01"
+  registration_confirmation_trigger = "registration-email-2026-01"
 }